RAG/pyproject.toml
[tool.poetry]
name = "rag"
version = "0.1.0"
description = ""
authors = ["Charles Wang <charleswang007@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.13"
langchain = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
pypdf = "^4.3.1"
langchain-qdrant = "0.2.0.dev1"
beautifulsoup4 = "^4.14.2"
lxml = "^6.0.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
